# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parcels tracking API",
        "contact": {
            "name": "API Support",
            "email": "bamboo-shop@example.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Localhost 8080"
        },
        {
            "url": "http://localhost:8000",
            "description": "Localhost 8000"
        }
    ],
    "paths": {
        "/api/login_check": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Operations about user",
                "operationId": "App\\Core\\Ports\\Rest\\AuthToken\\CreateAuthTokenAction::__invoke",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The username for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "apiKey",
                                            "example": "The token needed for requesting secure endpoints."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid username/password supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Page not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Page not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/parcels/{trackingId}": {
            "get": {
                "tags": [
                    "Parcels"
                ],
                "summary": "Find parcel by tracking ID",
                "operationId": "App\\Core\\Ports\\Rest\\Parcel\\GetParcelAction::__invoke",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tracking ID of parcel to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A parcel finded by the tracking ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Parcel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "JWT Token not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Page not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Page not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Parcel": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "trackingId": {
                        "type": "string"
                    },
                    "carrier": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "shippingDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    }
}